move Paligo API to paligo.go
